refactor: build site with rollup, share functions between scraper and site (#191)
