+++ lib/datetime.js
@@ -1,17 +1,17 @@
 /*
   Get a date object offset for the current timezone
 */
-function getDate() {
+export const getDate = function() {
   const date = new Date();
   const utcDate = new Date(date.toUTCString());
   utcDate.setHours(utcDate.getHours() - 7);
   return new Date(utcDate);
-}
+};
 
 /*
   Get date formatted in YYYY-M-D
 */
-function getYYYYMD(date = getDate(), sep = '-') {
+export const getYYYYMD = function(date = getDate(), sep = '-') {
   if (typeof date === 'string') {
     date = new Date(date);
   }
@@ -20,12 +20,12 @@ function getYYYYMD(date = getDate(), sep = '-') {
   const year = date.getUTCFullYear();
 
   return `${year}${sep}${month}${sep}${day}`;
-}
+};
 
 /*
   Get date formatted in YYYY-M-D
 */
-function getYYYYMMDD(date = getDate(), sep = '-') {
+export const getYYYYMMDD = function(date = getDate(), sep = '-') {
   if (typeof date === 'string') {
     date = new Date(date);
   }
@@ -37,12 +37,12 @@ function getYYYYMMDD(date = getDate(), sep = '-') {
   const year = date.getUTCFullYear();
 
   return `${year}${sep}${month}${sep}${day}`;
-}
+};
 
 /*
   Get date formatted in DD-MM-YYYY
 */
-function getDDMMYYYY(date = getDate(), sep = '-') {
+export const getDDMMYYYY = function(date = getDate(), sep = '-') {
   if (typeof date === 'string') {
     date = new Date(date);
   }
@@ -54,12 +54,12 @@ function getDDMMYYYY(date = getDate(), sep = '-') {
   const year = date.getUTCFullYear();
 
   return `${day}${sep}${month}${sep}${year}`;
-}
+};
 
 /*
   Get date formatted in M/D/YYYY
 */
-function getMDYYYY(date = getDate(), sep = '/') {
+export const getMDYYYY = function(date = getDate(), sep = '/') {
   if (typeof date === 'string') {
     date = new Date(date);
   }
@@ -68,12 +68,12 @@ function getMDYYYY(date = getDate(), sep = '/') {
   const year = date.getUTCFullYear();
 
   return `${month}${sep}${day}${sep}${year}`;
-}
+};
 
 /*
   Get date formatted in M/D/YY
 */
-function getMDYY(date = getDate(), sep = '/') {
+export const getMDYY = function(date = getDate(), sep = '/') {
   if (typeof date === 'string') {
     date = new Date(date);
   }
@@ -85,58 +85,58 @@ function getMDYY(date = getDate(), sep = '/') {
     .substr(2, 2);
 
   return `${month}${sep}${day}${sep}${year}`;
-}
+};
 
 /*
   Check of the *date* of the passed date is before the other passed date
   *sigh*
 */
-function dateIsBefore(a, b) {
+export const dateIsBefore = function(a, b) {
   a = new Date(a);
   b = new Date(b);
   a.setHours(0, 0, 0, 0);
   b.setHours(0, 0, 0, 0);
   return a.getTime() < b.getTime();
-}
+};
 
 /*
   Check of the *date* of the passed date is before or equal to the other passed date
   *sigh*
 */
-function dateIsBeforeOrEqualTo(a, b) {
+export const dateIsBeforeOrEqualTo = function(a, b) {
   a = new Date(a);
   b = new Date(b);
   a.setHours(0, 0, 0, 0);
   b.setHours(0, 0, 0, 0);
   return a.getTime() < b.getTime() || a.getTime() === b.getTime();
-}
+};
 
 /*
   Check if the date we're scraping is before the passed date
 */
-function scrapeDateIsBefore(date) {
+export const scrapeDateIsBefore = function(date) {
   let scrapeDate = getDate();
   if (process.env.SCRAPE_DATE) {
     scrapeDate = new Date(process.env.SCRAPE_DATE);
   }
   return dateIsBefore(scrapeDate, new Date(date));
-}
+};
 
 /*
   Check if the date we're scraping is after the passed date
 */
-function scrapeDateIsAfter(date) {
+export const scrapeDateIsAfter = function(date) {
   let scrapeDate = getDate();
   if (process.env.SCRAPE_DATE) {
     scrapeDate = new Date(process.env.SCRAPE_DATE);
   }
   return dateIsBefore(new Date(date), scrapeDate);
-}
+};
 
 /*
   Check if the date we're scraping is equal to the passed date
 */
-function scrapeDateIs(date) {
+export const scrapeDateIs = function(date) {
   let scrapeDate = getDate();
   if (process.env.SCRAPE_DATE) {
     scrapeDate = new Date(process.env.SCRAPE_DATE);
@@ -147,6 +147,4 @@ function scrapeDateIs(date) {
   compareDate.setHours(0, 0, 0, 0);
 
   return compareDate.getTime() === scrapeDate.getTime();
-}
-
-export { getDate, getYYYYMD, getYYYYMMDD, getMDYYYY, getMDYY, getDDMMYYYY, dateIsBefore, scrapeDateIsBefore, scrapeDateIsAfter, scrapeDateIs, dateIsBeforeOrEqualTo };
+};
