+++ lib/geography.test.js
@@ -1,14 +1,15 @@
 import each from 'jest-each';
 
-import * as transform from './transform.js';
+import * as geography from './geography.js';
 
-describe('Transform functions', () => {
+describe('Geography functions', () => {
   describe('toISO3166Alpha3', () => {
     const tests = [['France', 'FRA']];
     each(tests).test('when passed "%s", it returns %d', (str, expected) => {
-      expect(transform.toISO3166Alpha3(str)).toBe(expected);
+      expect(geography.toISO3166Alpha3(str)).toBe(expected);
     });
   });
+
   describe('addEmptyRegions', () => {
     test('does the doings', () => {
       let counties = [
@@ -18,7 +19,7 @@ describe('Transform functions', () => {
         }
       ];
       const countyNames = ['a', 'b', 'c'];
-      counties = transform.addEmptyRegions(counties, countyNames, 'county');
+      counties = geography.addEmptyRegions(counties, countyNames, 'county');
       expect(counties).toHaveLength(3);
       expect(counties[0].county).toEqual('a');
       expect(counties[1].county).toEqual('b');
