+++ lib/rules.js
@@ -5,7 +5,8 @@
   If the object has keys that match everything in at least one rejection criteria, it is rejected
   If the object has keys that match everything in at least one acceptance criteria, it's accepted
 */
-function isAcceptable(data, acceptCriteria, rejectCriteria) {
+// eslint-disable-next-line import/prefer-default-export
+export const isAcceptable = function(data, acceptCriteria, rejectCriteria) {
   // Fail things that match reject criteria
   if (rejectCriteria) {
     for (const [, criteria] of Object.entries(rejectCriteria)) {
@@ -35,7 +36,4 @@ function isAcceptable(data, acceptCriteria, rejectCriteria) {
   }
 
   return true;
-}
-
-// eslint-disable-next-line import/prefer-default-export
-export { isAcceptable };
+};
