+++ lib/stringify.js
@@ -1,7 +1,8 @@
 /*
   Generate a CSV from the given day
 */
-function csvForDay(data) {
+// eslint-disable-next-line import/prefer-default-export
+export const csvForDay = function(data) {
   // Start with the columns we want first
   let columns = ['city', 'county', 'state', 'country', 'cases', 'deaths', 'recovered', 'tested', 'active', 'population', 'lat', 'long', 'url'];
 
@@ -34,7 +35,4 @@ function csvForDay(data) {
     csvData.push(row);
   }
   return csvData;
-}
-
-// eslint-disable-next-line import/prefer-default-export
-export { csvForDay };
+};
