+++ scrapers/USA/AL/index.js
@@ -1,6 +1,7 @@
 import * as fetch from '../../../lib/fetch.js';
 import * as parse from '../../../lib/parse.js';
 import * as transform from '../../../lib/transform.js';
+import * as geography from '../../../lib/geography.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 // const UNASSIGNED = '(unassigned)';
@@ -90,7 +91,7 @@ const scraper = {
         return;
       }
       const $tr = $(tr);
-      const countyName = transform.addCounty(parse.string($tr.find('td:first-child').text()));
+      const countyName = geography.addCounty(parse.string($tr.find('td:first-child').text()));
       if (countyName === 'Out of State County') {
         return;
       }
@@ -99,7 +100,7 @@ const scraper = {
         cases: parse.number($tr.find('td:last-child').text())
       });
     });
-    counties = transform.addEmptyRegions(counties, this._counties, 'county');
+    counties = geography.addEmptyRegions(counties, this._counties, 'county');
 
     counties.push(transform.sumData(counties));
