+++ scrapers/USA/DE/index.js
@@ -2,6 +2,7 @@ import * as fetch from '../../../lib/fetch.js';
 import * as parse from '../../../lib/parse.js';
 import * as transform from '../../../lib/transform.js';
 import * as datetime from '../../../lib/datetime.js';
+import * as geography from '../../../lib/geography.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 // const UNASSIGNED = '(unassigned)';
@@ -24,7 +25,7 @@ const scraper = {
         .map(str => {
           const parts = str.split(': ');
           return {
-            county: transform.addCounty(parse.string(parts[0])),
+            county: geography.addCounty(parse.string(parts[0])),
             cases: parse.number(parts[1])
           };
         });
@@ -37,7 +38,7 @@ const scraper = {
     const counties = [];
     for (const county of data) {
       counties.push({
-        county: transform.addCounty(parse.string(county.NAME)),
+        county: geography.addCounty(parse.string(county.NAME)),
         cases: parse.number(county.Presumptive_Positive),
         recovered: parse.number(county.Recovered)
       });
