+++ scrapers/USA/FL/index.js
@@ -1,6 +1,7 @@
 import * as fetch from '../../../lib/fetch.js';
 import * as parse from '../../../lib/parse.js';
 import * as transform from '../../../lib/transform.js';
+import * as geography from '../../../lib/geography.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 const UNASSIGNED = '(unassigned)';
@@ -92,7 +93,7 @@ const scraper = {
           return;
         }
         const $tr = $(tr);
-        const county = transform.addCounty(parse.string($tr.find('td:nth-child(2)').text()));
+        const county = geography.addCounty(parse.string($tr.find('td:nth-child(2)').text()));
         countiesMap[county] = countiesMap[county] || { cases: 0 };
         countiesMap[county].cases += 1;
       });
@@ -110,7 +111,7 @@ const scraper = {
 
       counties.push(transform.sumData(counties));
 
-      counties = transform.addEmptyRegions(counties, this._counties, 'county');
+      counties = geography.addEmptyRegions(counties, this._counties, 'county');
 
       return counties;
     },
@@ -121,14 +122,14 @@ const scraper = {
       let counties = [];
       for (const county of data) {
         counties.push({
-          county: transform.addCounty(parse.string(county.County)),
+          county: geography.addCounty(parse.string(county.County)),
           cases: parse.number(county.Counts)
         });
       }
 
       counties.push(transform.sumData(counties));
 
-      counties = transform.addEmptyRegions(counties, this._counties, 'county');
+      counties = geography.addEmptyRegions(counties, this._counties, 'county');
 
       return counties;
     }
