+++ scrapers/USA/IA/index.js
@@ -1,6 +1,7 @@
 import * as fetch from '../../../lib/fetch.js';
 import * as parse from '../../../lib/parse.js';
 import * as transform from '../../../lib/transform.js';
+import * as geography from '../../../lib/geography.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 // const UNASSIGNED = '(unassigned)';
@@ -20,7 +21,7 @@ const scraper = {
       const $trs = $table.find('tbody > tr:not(:last-child)');
       $trs.each((index, tr) => {
         const $tr = $(tr);
-        const county = transform.addCounty(
+        const county = geography.addCounty(
           $tr
             .find('td:first-child')
             .text()
