+++ scrapers/USA/IL/index.js
@@ -1,6 +1,7 @@
 import * as fetch from '../../../lib/fetch.js';
 import * as parse from '../../../lib/parse.js';
 import * as transform from '../../../lib/transform.js';
+import * as geography from '../../../lib/geography.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 // const UNASSIGNED = '(unassigned)';
@@ -15,7 +16,7 @@ const scraper = {
     const counties = [];
     for (const county of data.characteristics_by_county.values) {
       counties.push({
-        county: transform.addCounty(county.County),
+        county: geography.addCounty(county.County),
         cases: parse.number(county.confirmed_cases),
         tested: parse.number(county.total_tested)
       });
