+++ scrapers/USA/NC/index.js
@@ -1,6 +1,7 @@
 import * as fetch from '../../../lib/fetch.js';
 import * as parse from '../../../lib/parse.js';
 import * as transform from '../../../lib/transform.js';
+import * as geography from '../../../lib/geography.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 // const UNASSIGNED = '(unassigned)';
@@ -15,7 +16,7 @@ const scraper = {
     const counties = [];
     for (const county of data) {
       counties.push({
-        county: transform.addCounty(parse.string(county.County)),
+        county: geography.addCounty(parse.string(county.County)),
         cases: parse.number(county.Total),
         deaths: parse.number(county.Deaths)
       });
