+++ scrapers/USA/NY/index.js
@@ -2,6 +2,7 @@ import * as fetch from '../../../lib/fetch.js';
 import * as parse from '../../../lib/parse.js';
 import * as transform from '../../../lib/transform.js';
 import * as datetime from '../../../lib/datetime.js';
+import * as geography from '../../../lib/geography.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 // const UNASSIGNED = '(unassigned)';
@@ -96,7 +97,7 @@ const scraper = {
       countyName = this._countyMap[countyName] || countyName;
       if (countyName !== 'New York State (Outside of NYC)' && countyName !== 'Total Positive Cases (Statewide)') {
         counties.push({
-          county: transform.addCounty(countyName),
+          county: geography.addCounty(countyName),
           cases: parse.number($tr.find('td:last-child').text())
         });
       }
@@ -104,7 +105,7 @@ const scraper = {
 
     counties.push(transform.sumData(counties));
 
-    counties = transform.addEmptyRegions(counties, this._counties, 'county');
+    counties = geography.addEmptyRegions(counties, this._counties, 'county');
 
     return counties;
   }
