+++ scrapers/USA/SD/index.js
@@ -1,6 +1,7 @@
 import * as fetch from '../../../lib/fetch.js';
 import * as parse from '../../../lib/parse.js';
 import * as transform from '../../../lib/transform.js';
+import * as geography from '../../../lib/geography.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 // const UNASSIGNED = '(unassigned)';
@@ -89,12 +90,12 @@ const scraper = {
       $trs.each((index, tr) => {
         const $tr = $(tr);
         counties.push({
-          county: transform.addCounty(parse.string($tr.find('> *:first-child').text())),
+          county: geography.addCounty(parse.string($tr.find('> *:first-child').text())),
           cases: parse.number($tr.find('> *:last-child').text())
         });
       });
       counties.push(transform.sumData(counties));
-      counties = transform.addEmptyRegions(counties, this._counties, 'county');
+      counties = geography.addEmptyRegions(counties, this._counties, 'county');
       return counties;
     },
     '2020-3-19': async function() {
@@ -108,12 +109,12 @@ const scraper = {
           return;
         }
         counties.push({
-          county: transform.addCounty(parse.string($tr.find('td:first-child').text())),
+          county: geography.addCounty(parse.string($tr.find('td:first-child').text())),
           cases: parse.number($tr.find('td:last-child').text())
         });
       });
       counties.push(transform.sumData(counties));
-      counties = transform.addEmptyRegions(counties, this._counties, 'county');
+      counties = geography.addEmptyRegions(counties, this._counties, 'county');
       return counties;
     }
   }
