+++ scrapers/USA/TN/index.js
@@ -1,6 +1,7 @@
 import * as fetch from '../../../lib/fetch.js';
 import * as parse from '../../../lib/parse.js';
 import * as transform from '../../../lib/transform.js';
+import * as geography from '../../../lib/geography.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 // const UNASSIGNED = '(unassigned)';
@@ -119,14 +120,14 @@ const scraper = {
       }
       const $tr = $(tr);
       counties.push({
-        county: transform.addCounty(parse.string($tr.find('td:first-child').text())),
+        county: geography.addCounty(parse.string($tr.find('td:first-child').text())),
         cases: parse.number($tr.find('td:last-child').text())
       });
     });
 
     counties.push(transform.sumData(counties));
 
-    counties = transform.addEmptyRegions(counties, this._counties, 'county');
+    counties = geography.addEmptyRegions(counties, this._counties, 'county');
 
     return counties;
   }
