+++ scrapers/USA/TX/index.js
@@ -2,6 +2,7 @@ import * as fetch from '../../../lib/fetch.js';
 import * as parse from '../../../lib/parse.js';
 import * as transform from '../../../lib/transform.js';
 import * as datetime from '../../../lib/datetime.js';
+import * as geography from '../../../lib/geography.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 // const UNASSIGNED = '(unassigned)';
@@ -282,7 +283,7 @@ const scraper = {
     const $trs = $table.find('tbody > tr:not(:last-child)');
     $trs.each((index, tr) => {
       const $tr = $(tr);
-      const county = transform.addCounty(
+      const county = geography.addCounty(
         $tr
           .find('td:first-child')
           .text()
@@ -294,7 +295,7 @@ const scraper = {
         cases
       });
     });
-    counties = transform.addEmptyRegions(counties, this._counties, 'county');
+    counties = geography.addEmptyRegions(counties, this._counties, 'county');
     counties.push(transform.sumData(counties));
     return counties;
   }
