+++ scrapers/USA/USA-CDC.js
@@ -1,6 +1,6 @@
 import * as fetch from '../../lib/fetch.js';
 import * as parse from '../../lib/parse.js';
-import * as transform from '../../lib/transform.js';
+import * as geography from '../../lib/geography.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 // const UNASSIGNED = '(unassigned)';
@@ -27,7 +27,7 @@ const scraper = {
     for (const stateData of data) {
       if (stateData.Name) {
         states.push({
-          state: transform.toUSStateAbbreviation(parse.string(stateData.Name)),
+          state: geography.toUSStateAbbreviation(parse.string(stateData.Name)),
           cases: this._getCaseNumber(stateData['Cases Reported'])
         });
       }
