+++ scrapers/USA/WY/index.js
@@ -3,6 +3,7 @@ import * as fetch from '../../../lib/fetch.js';
 import * as parse from '../../../lib/parse.js';
 import * as transform from '../../../lib/transform.js';
 import maintainers from '../../../lib/maintainers.js';
+import * as geography from '../../../lib/geography.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 // const UNASSIGNED = '(unassigned)';
@@ -27,7 +28,7 @@ const scraper = {
         continue;
       }
       counties.push({
-        county: transform.addCounty(parse.string(county)),
+        county: geography.addCounty(parse.string(county)),
         cases: parse.number(count || 0)
       });
     }
