+++ site/lib/color.js
@@ -1,0 +1,6 @@
+import * as d3color from 'd3-color';
+
+// eslint-disable-next-line import/prefer-default-export
+export const getLightness = function(c) {
+  return d3color.lab(c).l;
+};
